fix(admin): Incorrect User index after updating
Fixes an issue where the user index defaults to -1 when the first user
in the list (index: `0`) is edited. The result of the bug is that a
duplicate user is added to the first position in the list.

The fix is to do an equality check on index for when the index is `0`,
rather than checking for it's existence.