fix(activists): Activist list missing status field
Fixes an issue with the Activists list not showing the Status field upon
initial load.

The issues was that the "status" field was not getting calculated in the
getActivistRange method of the activist model.

The fix was to iterate through the list and call the getStatus method on
each activist before returning the list.