fix(admin): User create bug
Fixes a bug with the Admin's User create.

The issue was caused by making a reference to the current user list based on the userIndex which is -1 when creating a new User, so the user could not be found to update the Roles list.

Currently, we're not adding Roles when creating a new User so there's no need to update the property.