add activists.pronouns column
This commit adds a `pronouns` column to the `activists` table. This
includes additional wiring for all of the CRUD operations relating to
it, copy/pasted from handling of the `phone` column.

Notably, there's already a 20-character `pronouns` column in the
`form_application` table; evidently, it originally existed but was
removed in mid 2019. For simplicity, I kept the 20-character limit.

Per Cassie's request, this is a free-form text box.

Existing ADB SQL databases can be migrated by running:

```
ALTER TABLE activists ADD pronouns VARCHAR(20) NOT NULL DEFAULT '';
```