Fix misplaced City dropdown on dxe.io/join
Previously, the City field autocomplete dropdown would be positioned
correctly only if the user did not scroll from the top of the page, a
known issue with vue-google-autocomplete:
https://github.com/MadimetjaShika/vuetify-google-autocomplete/issues/100

On small screens, if the "City" field was "below the fold", the dropdown
would be hidden entirely.

The issue is fixed by removing "overflow: auto" on html/body, which was
added here:
https://github.com/dxe/adb/commit/ae799dcf1795c20fed7a3dd3aefd60bfc9db5542#diff-a5ea33c888430601a659bcbef2da1944097953737f2606282efc13ca6e5fbabaR2

After removing overflow: auto, white margins appeared above and below
the background image. Removing the duplicate body CSS rule fixed this.
Removing "height: 100%" fixed something else that started happening too.
As far as I can tell, the app works the same as before with all of
these changes combined.

Not sure if a new version of vue-google-autocomplete would fix it but
anyway we can't upgrade to the latest version because it requires
upgrading Vue to at least 2.7.