Increase size limit for Facebook token column
The database and UI had a 200 character limit, but Facebook generated a
token we need to use that has 203 characters. This increases the limit
to 500 so we hopefully we won't need to worry about future changes to
the size. The prod database has already been manually updated:

`alter table fb_pages modify token varchar(500);`