chore(forms): remove redundant duplicate checks
Multi-chapter support would require adding even more complexity to these
redundant SQL checks.

The current logic for interest/application form processing is:
* Check for existing activist with same name and update it
* If no match, check for existing activist with same email and update it
* If no match, insert new activist

The SQL query for inserting a new activist checks that there is no
activist with the same name or email, but that is handled by the Go code
prior to executing the SQL query.

This change removes the redundant checks in the insert query.

The aforementioned logic does not all run in one transaction, but a race
condition here is unlikely and the consequences are acceptable.