feat: add timestamp columns to activists
This allows smarter merging of activist records. This is especially
important because in the future we may decide to merge activist records
across chapters. We suspect there may be a lot of activists in multiple
chapters now that check-in forms systematically duplicate activists
across chapters when someone gives a zip code near one chapter (which is
presumed to be their home chapter) on a check-in form belonging to
another chapter (which needs the same activist info for attendance).