feat: add timestamp columns to activists (#274)
* feat: add timestamp columns to activists

This allows smarter merging of activist records. This is especially
important because in the future we may decide to merge activist records
across chapters. We suspect there may be a lot of activists in multiple
chapters now that check-in forms systematically duplicate activists
across chapters when someone gives a zip code near one chapter (which is
presumed to be their home chapter) on a check-in form belonging to
another chapter (which needs the same activist info for attendance).

* chore: deduplicate update queries for activist update. One was missing
dev_application_date and dev_application_type and the other was missing lat, lng,
maybe on purpose, but it was complicated. Combining so all fields will get
updated in all cases.

* chore: add unit tests for inserting, updating and merging activists.
Move activist test utils to avoid cyclic dependency

* chore: deduplicate form processor sql logic and fix one inconsistency where we were
checking that the name is not blank in a case where name is irrelevant because it's
matching on email instead.