Frontend-v2: re-initialize shadcn
Shadcn recently added support for Tanstack forms. Before adding our
first usage of this feature, this change more or less removes our shadcn
usage and re-adds it from scratch in order to get updates not just to
dependencies like radix-ui components, but also updates to the
components themselves. Updating components themselves usually isn't
necessary, but sometimes it is (such as was needed recently to upgrade
to Tailwind 4). Not many customizations needed to be reapplied so it
was easy to change now, just in case.