fix(eventbrite): continue syncing in face of errors
When one event fails to sync, all subsequent ones for the chapter are
not synced. This is making it difficult to discover patterns with
why some events do not sync, and anyway is somewhat arbitrary in what
events are synced. This could cause issues with too many failed requests
but let's hope it's fine.